Domain Search: Consolidate free subdomain helpers into single file
Combine is-wpcom-subdomain-query.ts and is-blog-subdomain-query.ts into
is-free-subdomain-query.ts and extract isFreeSubdomainQuery helper to
simplify consumers.
