+++ packages/domain-search/src/helpers/index.ts
@@ -1,7 +1,11 @@
 export { getRootDomain } from './get-root-domain';
 export { getTld } from './get-tld';
 export { isSubdomain } from './is-subdomain';
-export { isBlogSubdomainQuery } from './is-blog-subdomain-query';
-export { isWpcomSubdomainQuery, stripWpcomSubdomainSuffix } from './is-wpcom-subdomain-query';
+export {
+	isBlogSubdomainQuery,
+	isFreeSubdomainQuery,
+	isWpcomSubdomainQuery,
+	stripWpcomSubdomainSuffix,
+} from './is-free-subdomain-query';
 export { parseDomainAgainstTldList } from './parse-domain-against-tld-list';
 export { parseMatchReasons } from './parse-match-reasons';
