+++ packages/domain-search/src/helpers/is-free-subdomain-query.ts
@@ -1,3 +1,5 @@
+const WPCOM_SUBDOMAIN_SUFFIX = '.wordpress.com';
+
 /**
  * List of free .blog subdomains offered by WordPress.com.
  * Duplicated from packages/domains-table/src/utils/is-free-url-domain-name.ts
@@ -40,9 +42,36 @@ const BLOG_SUBDOMAIN_SUFFIXES = FREE_DOT_BLOG_SUBDOMAINS.map(
 );
 
 /**
+ * Detects whether a search query is for a WordPress.com subdomain
+ * (e.g. "mysite.wordpress.com").
+ */
+export function isWpcomSubdomainQuery( query: string ): boolean {
+	return query.endsWith( WPCOM_SUBDOMAIN_SUFFIX );
+}
+
+/**
+ * Strips the WordPress.com subdomain suffix from a query, returning just the
+ * subdomain label (e.g. "mysite.wordpress.com" → "mysite").
+ */
+export function stripWpcomSubdomainSuffix( query: string ): string {
+	if ( query.endsWith( WPCOM_SUBDOMAIN_SUFFIX ) ) {
+		return query.slice( 0, -WPCOM_SUBDOMAIN_SUFFIX.length );
+	}
+	return query;
+}
+
+/**
  * Detects whether a search query is for a free .blog subdomain
  * (e.g. "test.tech.blog", "mysite.photo.blog").
  */
 export function isBlogSubdomainQuery( query: string ): boolean {
 	return BLOG_SUBDOMAIN_SUFFIXES.some( ( suffix ) => query.endsWith( suffix ) );
+}
+
+/**
+ * Detects whether a search query is for any free subdomain
+ * (.wordpress.com or free .blog subdomains).
+ */
+export function isFreeSubdomainQuery( query: string ): boolean {
+	return isWpcomSubdomainQuery( query ) || isBlogSubdomainQuery( query );
 }
